Feat!: completed milestone 1

+++ src/main/java/in/fssa/homebakery/servlets/DeleteProductPriceServlet.java
@@ -12,13 +12,14 @@ import in.fssa.homebakery.service.ProductPriceService;
 /**
  * Servlet implementation class DeleteProductPriceServlet
  */
-@WebServlet("/product/price/delete")
+@WebServlet("/admin/product/price/delete")
 public class DeleteProductPriceServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 	
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		
 		String paramId = request.getParameter("priceid");
+		String paramProdId = request.getParameter("productid");
 	
 		ProductPriceService productPriceService = new ProductPriceService();
 		
@@ -30,11 +31,21 @@ public class DeleteProductPriceServlet extends HttpServlet {
 			throw new RuntimeException("Id cannnot be empty");
 		}
 		
+		if(paramProdId == null) {
+			throw new RuntimeException("Id cannnot be null");
+		}
+		
+		if("".equals(paramProdId)) {
+			throw new RuntimeException("Id cannnot be empty");
+		}
+		
 		try {
 			int id = Integer.parseInt(paramId);
+			int prodId = Integer.parseInt(paramProdId);
 			
 			productPriceService.deleteProductPrice(id);
-			response.sendRedirect(request.getContextPath() + "/products");
+			
+			response.sendRedirect(request.getContextPath() + "/admin/product?productid=" + prodId);
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
